// src/app/core/enums/lp.ts
export enum ChipsLpType {
  Wbnb = 'WBNB',
  Busd = 'BUSD',
  Btd = 'BTD',
  Mdo = 'MDO',
  ChipsWbnb = 'CHIPS_WBNB'
}

export enum FishLpV1Type {
  ChipsBusd = 'CHIPS_BUSD',
  ChipsWbnb = 'CHIPS_WBNB',
  FishWbnb = 'FISH_WBNB',
  ChipsBtd = 'CHIPS_BTD'
}

export enum FishLpV2Type {
  ChipsBusd = 'CHIPS_BUSD',
  ChipsWbnb = 'CHIPS_WBNB',
  FishWbnb = 'FISH_WBNB',
  ChipsBtd = 'CHIPS_BTD',
  MpeasChips = 'MPEAS_CHIPS'
}

export enum LPActionType {
  Deposit = 'DEPOSIT',
  Withdraw = 'WITHDRAW',
  Claim = 'CLAIM'
}

export enum FarmType {
  Chips = 'CHIPS',
  Fish = 'FISH'
}